// src/ts/src/generated/semantic-surface-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

export enum SemanticSurfaceType {
  RoofSurface = 0,
  GroundSurface = 1,
  WallSurface = 2,
  ClosureSurface = 3,
  OuterCeilingSurface = 4,
  OuterFloorSurface = 5,
  Window = 6,
  Door = 7,
  InteriorWallSurface = 8,
  CeilingSurface = 9,
  FloorSurface = 10,
  WaterSurface = 11,
  WaterGroundSurface = 12,
  WaterClosureSurface = 13,
  TrafficArea = 14,
  AuxiliaryTrafficArea = 15,
  TransportationMarking = 16,
  TransportationHole = 17,
  ExtraSemanticSurface = 18
}